test: separate fixture object to TestFixture.kt
