+++ src/test/kotlin/blackjack/model/holder/HandTest.kt
@@ -1,0 +1,10 @@
+package blackjack.model.holder
+
+import org.junit.jupiter.api.Test
+
+class HandTest {
+    @Test
+    fun `init Hand`() {
+        Hand()
+    }
+}
